Rename `authentification` to `authentication`.
