+++ bot/src/authentication.ts
